move properties file edit to support
